tests for service code generation in sep files
Splits out the giant 3500+ line `pkg/model/crd_test.go` file into
separate service-specific test files. This should be easier to deal with
in the long-term instead of trying to navigate a super-long single file.

Issue #212
