+++ pkg/model/dynamodb_test.go
@@ -1,0 +1,445 @@
+// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	 http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package model_test
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/aws/aws-controllers-k8s/pkg/model"
+	"github.com/aws/aws-controllers-k8s/pkg/testutil"
+)
+
+func TestDynamoDB_Table(t *testing.T) {
+	assert := assert.New(t)
+	require := require.New(t)
+
+	sh := testutil.NewSchemaHelperForService(t, "dynamodb")
+
+	crds, err := sh.GetCRDs()
+	require.Nil(err)
+
+	crd := getCRDByName("Table", crds)
+	require.NotNil(crd)
+
+	// The DynamoDB Table API has these operations:
+	//
+	// * CreateTable
+	// * DeleteTable
+	// * DescribeTable
+	// * ListTables
+	// * UpdateTable
+	require.NotNil(crd.Ops)
+
+	assert.NotNil(crd.Ops.Create)
+	assert.NotNil(crd.Ops.Delete)
+	assert.NotNil(crd.Ops.ReadOne)
+	assert.NotNil(crd.Ops.ReadMany)
+	assert.NotNil(crd.Ops.Update)
+
+	assert.Nil(crd.Ops.GetAttributes)
+	assert.Nil(crd.Ops.SetAttributes)
+
+	// The DescribeTable operation has the following definition:
+	//
+	//    "DescribeTable":{
+	//      "name":"DescribeTable",
+	//      "http":{
+	//        "method":"POST",
+	//        "requestUri":"/"
+	//      },
+	//      "input":{"shape":"DescribeTableInput"},
+	//      "output":{"shape":"DescribeTableOutput"},
+	//      "errors":[
+	//        {"shape":"ResourceNotFoundException"},
+	//        {"shape":"InternalServerError"}
+	//      ],
+	//      "endpointdiscovery":{
+	//      }
+	//    },
+	//
+	// Where the ResourceNotFoundException shape looks like this:
+	//
+	//    "ResourceNotFoundException":{
+	//      "type":"structure",
+	//      "members":{
+	//        "message":{"shape":"ErrorMessage"}
+	//      },
+	//      "exception":true
+	//    },
+	//
+	//
+	// Which does not indicate that the error is a 404 :( So, the logic in the
+	// CRD.ExceptionCode(404) method needs to get its override from the
+	// generate.yaml configuration file.
+	assert.Equal("ResourceNotFoundException", crd.ExceptionCode(404))
+
+	specFields := crd.SpecFields
+	statusFields := crd.StatusFields
+
+	expSpecFieldCamel := []string{
+		"AttributeDefinitions",
+		"BillingMode",
+		"GlobalSecondaryIndexes",
+		"KeySchema",
+		"LocalSecondaryIndexes",
+		"ProvisionedThroughput",
+		"SSESpecification",
+		"StreamSpecification",
+		"TableName",
+		"Tags",
+	}
+	assert.Equal(expSpecFieldCamel, attrCamelNames(specFields))
+
+	expCreateInput := `
+	if r.ko.Spec.AttributeDefinitions != nil {
+		f0 := []*svcsdk.AttributeDefinition{}
+		for _, f0iter := range r.ko.Spec.AttributeDefinitions {
+			f0elem := &svcsdk.AttributeDefinition{}
+			if f0iter.AttributeName != nil {
+				f0elem.SetAttributeName(*f0iter.AttributeName)
+			}
+			if f0iter.AttributeType != nil {
+				f0elem.SetAttributeType(*f0iter.AttributeType)
+			}
+			f0 = append(f0, f0elem)
+		}
+		res.SetAttributeDefinitions(f0)
+	}
+	if r.ko.Spec.BillingMode != nil {
+		res.SetBillingMode(*r.ko.Spec.BillingMode)
+	}
+	if r.ko.Spec.GlobalSecondaryIndexes != nil {
+		f2 := []*svcsdk.GlobalSecondaryIndex{}
+		for _, f2iter := range r.ko.Spec.GlobalSecondaryIndexes {
+			f2elem := &svcsdk.GlobalSecondaryIndex{}
+			if f2iter.IndexName != nil {
+				f2elem.SetIndexName(*f2iter.IndexName)
+			}
+			if f2iter.KeySchema != nil {
+				f2elemf1 := []*svcsdk.KeySchemaElement{}
+				for _, f2elemf1iter := range f2iter.KeySchema {
+					f2elemf1elem := &svcsdk.KeySchemaElement{}
+					if f2elemf1iter.AttributeName != nil {
+						f2elemf1elem.SetAttributeName(*f2elemf1iter.AttributeName)
+					}
+					if f2elemf1iter.KeyType != nil {
+						f2elemf1elem.SetKeyType(*f2elemf1iter.KeyType)
+					}
+					f2elemf1 = append(f2elemf1, f2elemf1elem)
+				}
+				f2elem.SetKeySchema(f2elemf1)
+			}
+			if f2iter.Projection != nil {
+				f2elemf2 := &svcsdk.Projection{}
+				if f2iter.Projection.NonKeyAttributes != nil {
+					f2elemf2f0 := []*string{}
+					for _, f2elemf2f0iter := range f2iter.Projection.NonKeyAttributes {
+						var f2elemf2f0elem string
+						f2elemf2f0elem = *f2elemf2f0iter
+						f2elemf2f0 = append(f2elemf2f0, &f2elemf2f0elem)
+					}
+					f2elemf2.SetNonKeyAttributes(f2elemf2f0)
+				}
+				if f2iter.Projection.ProjectionType != nil {
+					f2elemf2.SetProjectionType(*f2iter.Projection.ProjectionType)
+				}
+				f2elem.SetProjection(f2elemf2)
+			}
+			if f2iter.ProvisionedThroughput != nil {
+				f2elemf3 := &svcsdk.ProvisionedThroughput{}
+				if f2iter.ProvisionedThroughput.ReadCapacityUnits != nil {
+					f2elemf3.SetReadCapacityUnits(*f2iter.ProvisionedThroughput.ReadCapacityUnits)
+				}
+				if f2iter.ProvisionedThroughput.WriteCapacityUnits != nil {
+					f2elemf3.SetWriteCapacityUnits(*f2iter.ProvisionedThroughput.WriteCapacityUnits)
+				}
+				f2elem.SetProvisionedThroughput(f2elemf3)
+			}
+			f2 = append(f2, f2elem)
+		}
+		res.SetGlobalSecondaryIndexes(f2)
+	}
+	if r.ko.Spec.KeySchema != nil {
+		f3 := []*svcsdk.KeySchemaElement{}
+		for _, f3iter := range r.ko.Spec.KeySchema {
+			f3elem := &svcsdk.KeySchemaElement{}
+			if f3iter.AttributeName != nil {
+				f3elem.SetAttributeName(*f3iter.AttributeName)
+			}
+			if f3iter.KeyType != nil {
+				f3elem.SetKeyType(*f3iter.KeyType)
+			}
+			f3 = append(f3, f3elem)
+		}
+		res.SetKeySchema(f3)
+	}
+	if r.ko.Spec.LocalSecondaryIndexes != nil {
+		f4 := []*svcsdk.LocalSecondaryIndex{}
+		for _, f4iter := range r.ko.Spec.LocalSecondaryIndexes {
+			f4elem := &svcsdk.LocalSecondaryIndex{}
+			if f4iter.IndexName != nil {
+				f4elem.SetIndexName(*f4iter.IndexName)
+			}
+			if f4iter.KeySchema != nil {
+				f4elemf1 := []*svcsdk.KeySchemaElement{}
+				for _, f4elemf1iter := range f4iter.KeySchema {
+					f4elemf1elem := &svcsdk.KeySchemaElement{}
+					if f4elemf1iter.AttributeName != nil {
+						f4elemf1elem.SetAttributeName(*f4elemf1iter.AttributeName)
+					}
+					if f4elemf1iter.KeyType != nil {
+						f4elemf1elem.SetKeyType(*f4elemf1iter.KeyType)
+					}
+					f4elemf1 = append(f4elemf1, f4elemf1elem)
+				}
+				f4elem.SetKeySchema(f4elemf1)
+			}
+			if f4iter.Projection != nil {
+				f4elemf2 := &svcsdk.Projection{}
+				if f4iter.Projection.NonKeyAttributes != nil {
+					f4elemf2f0 := []*string{}
+					for _, f4elemf2f0iter := range f4iter.Projection.NonKeyAttributes {
+						var f4elemf2f0elem string
+						f4elemf2f0elem = *f4elemf2f0iter
+						f4elemf2f0 = append(f4elemf2f0, &f4elemf2f0elem)
+					}
+					f4elemf2.SetNonKeyAttributes(f4elemf2f0)
+				}
+				if f4iter.Projection.ProjectionType != nil {
+					f4elemf2.SetProjectionType(*f4iter.Projection.ProjectionType)
+				}
+				f4elem.SetProjection(f4elemf2)
+			}
+			f4 = append(f4, f4elem)
+		}
+		res.SetLocalSecondaryIndexes(f4)
+	}
+	if r.ko.Spec.ProvisionedThroughput != nil {
+		f5 := &svcsdk.ProvisionedThroughput{}
+		if r.ko.Spec.ProvisionedThroughput.ReadCapacityUnits != nil {
+			f5.SetReadCapacityUnits(*r.ko.Spec.ProvisionedThroughput.ReadCapacityUnits)
+		}
+		if r.ko.Spec.ProvisionedThroughput.WriteCapacityUnits != nil {
+			f5.SetWriteCapacityUnits(*r.ko.Spec.ProvisionedThroughput.WriteCapacityUnits)
+		}
+		res.SetProvisionedThroughput(f5)
+	}
+	if r.ko.Spec.SSESpecification != nil {
+		f6 := &svcsdk.SSESpecification{}
+		if r.ko.Spec.SSESpecification.Enabled != nil {
+			f6.SetEnabled(*r.ko.Spec.SSESpecification.Enabled)
+		}
+		if r.ko.Spec.SSESpecification.KMSMasterKeyID != nil {
+			f6.SetKMSMasterKeyId(*r.ko.Spec.SSESpecification.KMSMasterKeyID)
+		}
+		if r.ko.Spec.SSESpecification.SSEType != nil {
+			f6.SetSSEType(*r.ko.Spec.SSESpecification.SSEType)
+		}
+		res.SetSSESpecification(f6)
+	}
+	if r.ko.Spec.StreamSpecification != nil {
+		f7 := &svcsdk.StreamSpecification{}
+		if r.ko.Spec.StreamSpecification.StreamEnabled != nil {
+			f7.SetStreamEnabled(*r.ko.Spec.StreamSpecification.StreamEnabled)
+		}
+		if r.ko.Spec.StreamSpecification.StreamViewType != nil {
+			f7.SetStreamViewType(*r.ko.Spec.StreamSpecification.StreamViewType)
+		}
+		res.SetStreamSpecification(f7)
+	}
+	if r.ko.Spec.TableName != nil {
+		res.SetTableName(*r.ko.Spec.TableName)
+	}
+	if r.ko.Spec.Tags != nil {
+		f9 := []*svcsdk.Tag{}
+		for _, f9iter := range r.ko.Spec.Tags {
+			f9elem := &svcsdk.Tag{}
+			if f9iter.Key != nil {
+				f9elem.SetKey(*f9iter.Key)
+			}
+			if f9iter.Value != nil {
+				f9elem.SetValue(*f9iter.Value)
+			}
+			f9 = append(f9, f9elem)
+		}
+		res.SetTags(f9)
+	}
+`
+	assert.Equal(expCreateInput, crd.GoCodeSetInput(model.OpTypeCreate, "r.ko", "res", 1))
+
+	expStatusFieldCamel := []string{
+		"ArchivalSummary",
+		"BillingModeSummary",
+		"CreationDateTime",
+		"GlobalTableVersion",
+		"ItemCount",
+		"LatestStreamARN",
+		"LatestStreamLabel",
+		"Replicas",
+		"RestoreSummary",
+		"SSEDescription",
+		"TableID",
+		"TableSizeBytes",
+		"TableStatus",
+	}
+	assert.Equal(expStatusFieldCamel, attrCamelNames(statusFields))
+
+	// The DynamoDB API uses an API that uses "wrapper" single-member objects
+	// in the JSON response for the create/describe calls. In other words, the
+	// returned result from the DescribeTable API looks like this:
+	//
+	// {
+	//   "table": {
+	//	 .. bunch of fields for the table ..
+	//   }
+	// }
+	//
+	// However, the *ShapeName* of the "table" field is actually
+	// TableDescription. This tests that we're properly outputting the
+	// memberName (which is "Table" and not "TableDescription") when we build
+	// the Table CRD's Status field from the DescribeTableOutput shape.
+	expReadOneOutput := `
+	if resp.Table.ArchivalSummary != nil {
+		f0 := &svcapitypes.ArchivalSummary{}
+		if resp.Table.ArchivalSummary.ArchivalBackupArn != nil {
+			f0.ArchivalBackupARN = resp.Table.ArchivalSummary.ArchivalBackupArn
+		}
+		if resp.Table.ArchivalSummary.ArchivalDateTime != nil {
+			f0.ArchivalDateTime = &metav1.Time{*resp.Table.ArchivalSummary.ArchivalDateTime}
+		}
+		if resp.Table.ArchivalSummary.ArchivalReason != nil {
+			f0.ArchivalReason = resp.Table.ArchivalSummary.ArchivalReason
+		}
+		ko.Status.ArchivalSummary = f0
+	}
+	if resp.Table.BillingModeSummary != nil {
+		f2 := &svcapitypes.BillingModeSummary{}
+		if resp.Table.BillingModeSummary.BillingMode != nil {
+			f2.BillingMode = resp.Table.BillingModeSummary.BillingMode
+		}
+		if resp.Table.BillingModeSummary.LastUpdateToPayPerRequestDateTime != nil {
+			f2.LastUpdateToPayPerRequestDateTime = &metav1.Time{*resp.Table.BillingModeSummary.LastUpdateToPayPerRequestDateTime}
+		}
+		ko.Status.BillingModeSummary = f2
+	}
+	if resp.Table.CreationDateTime != nil {
+		ko.Status.CreationDateTime = &metav1.Time{*resp.Table.CreationDateTime}
+	}
+	if resp.Table.GlobalTableVersion != nil {
+		ko.Status.GlobalTableVersion = resp.Table.GlobalTableVersion
+	}
+	if resp.Table.ItemCount != nil {
+		ko.Status.ItemCount = resp.Table.ItemCount
+	}
+	if resp.Table.LatestStreamArn != nil {
+		ko.Status.LatestStreamARN = resp.Table.LatestStreamArn
+	}
+	if resp.Table.LatestStreamLabel != nil {
+		ko.Status.LatestStreamLabel = resp.Table.LatestStreamLabel
+	}
+	if resp.Table.Replicas != nil {
+		f12 := []*svcapitypes.ReplicaDescription{}
+		for _, f12iter := range resp.Table.Replicas {
+			f12elem := &svcapitypes.ReplicaDescription{}
+			if f12iter.GlobalSecondaryIndexes != nil {
+				f12elemf0 := []*svcapitypes.ReplicaGlobalSecondaryIndexDescription{}
+				for _, f12elemf0iter := range f12iter.GlobalSecondaryIndexes {
+					f12elemf0elem := &svcapitypes.ReplicaGlobalSecondaryIndexDescription{}
+					if f12elemf0iter.IndexName != nil {
+						f12elemf0elem.IndexName = f12elemf0iter.IndexName
+					}
+					if f12elemf0iter.ProvisionedThroughputOverride != nil {
+						f12elemf0elemf1 := &svcapitypes.ProvisionedThroughputOverride{}
+						if f12elemf0iter.ProvisionedThroughputOverride.ReadCapacityUnits != nil {
+							f12elemf0elemf1.ReadCapacityUnits = f12elemf0iter.ProvisionedThroughputOverride.ReadCapacityUnits
+						}
+						f12elemf0elem.ProvisionedThroughputOverride = f12elemf0elemf1
+					}
+					f12elemf0 = append(f12elemf0, f12elemf0elem)
+				}
+				f12elem.GlobalSecondaryIndexes = f12elemf0
+			}
+			if f12iter.KMSMasterKeyId != nil {
+				f12elem.KMSMasterKeyID = f12iter.KMSMasterKeyId
+			}
+			if f12iter.ProvisionedThroughputOverride != nil {
+				f12elemf2 := &svcapitypes.ProvisionedThroughputOverride{}
+				if f12iter.ProvisionedThroughputOverride.ReadCapacityUnits != nil {
+					f12elemf2.ReadCapacityUnits = f12iter.ProvisionedThroughputOverride.ReadCapacityUnits
+				}
+				f12elem.ProvisionedThroughputOverride = f12elemf2
+			}
+			if f12iter.RegionName != nil {
+				f12elem.RegionName = f12iter.RegionName
+			}
+			if f12iter.ReplicaStatus != nil {
+				f12elem.ReplicaStatus = f12iter.ReplicaStatus
+			}
+			if f12iter.ReplicaStatusDescription != nil {
+				f12elem.ReplicaStatusDescription = f12iter.ReplicaStatusDescription
+			}
+			if f12iter.ReplicaStatusPercentProgress != nil {
+				f12elem.ReplicaStatusPercentProgress = f12iter.ReplicaStatusPercentProgress
+			}
+			f12 = append(f12, f12elem)
+		}
+		ko.Status.Replicas = f12
+	}
+	if resp.Table.RestoreSummary != nil {
+		f13 := &svcapitypes.RestoreSummary{}
+		if resp.Table.RestoreSummary.RestoreDateTime != nil {
+			f13.RestoreDateTime = &metav1.Time{*resp.Table.RestoreSummary.RestoreDateTime}
+		}
+		if resp.Table.RestoreSummary.RestoreInProgress != nil {
+			f13.RestoreInProgress = resp.Table.RestoreSummary.RestoreInProgress
+		}
+		if resp.Table.RestoreSummary.SourceBackupArn != nil {
+			f13.SourceBackupARN = resp.Table.RestoreSummary.SourceBackupArn
+		}
+		if resp.Table.RestoreSummary.SourceTableArn != nil {
+			f13.SourceTableARN = resp.Table.RestoreSummary.SourceTableArn
+		}
+		ko.Status.RestoreSummary = f13
+	}
+	if resp.Table.SSEDescription != nil {
+		f14 := &svcapitypes.SSEDescription{}
+		if resp.Table.SSEDescription.InaccessibleEncryptionDateTime != nil {
+			f14.InaccessibleEncryptionDateTime = &metav1.Time{*resp.Table.SSEDescription.InaccessibleEncryptionDateTime}
+		}
+		if resp.Table.SSEDescription.KMSMasterKeyArn != nil {
+			f14.KMSMasterKeyARN = resp.Table.SSEDescription.KMSMasterKeyArn
+		}
+		if resp.Table.SSEDescription.SSEType != nil {
+			f14.SSEType = resp.Table.SSEDescription.SSEType
+		}
+		if resp.Table.SSEDescription.Status != nil {
+			f14.Status = resp.Table.SSEDescription.Status
+		}
+		ko.Status.SSEDescription = f14
+	}
+	if resp.Table.TableId != nil {
+		ko.Status.TableID = resp.Table.TableId
+	}
+	if resp.Table.TableSizeBytes != nil {
+		ko.Status.TableSizeBytes = resp.Table.TableSizeBytes
+	}
+	if resp.Table.TableStatus != nil {
+		ko.Status.TableStatus = resp.Table.TableStatus
+	}
+`
+	assert.Equal(expReadOneOutput, crd.GoCodeSetOutput(model.OpTypeGet, "resp", "ko.Status", 1))
+}
