+++ pkg/model/ec2_test.go
@@ -1,0 +1,531 @@
+// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	 http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package model_test
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/aws/aws-controllers-k8s/pkg/model"
+	"github.com/aws/aws-controllers-k8s/pkg/testutil"
+)
+
+func TestEC2_LaunchTemplate(t *testing.T) {
+	assert := assert.New(t)
+	require := require.New(t)
+
+	sh := testutil.NewSchemaHelperForService(t, "ec2")
+
+	crds, err := sh.GetCRDs()
+	require.Nil(err)
+
+	crd := getCRDByName("LaunchTemplate", crds)
+	require.NotNil(crd)
+
+	assert.Equal("LaunchTemplate", crd.Names.Camel)
+	assert.Equal("launchTemplate", crd.Names.CamelLower)
+	assert.Equal("launch_template", crd.Names.Snake)
+
+	// The DescribeLaunchTemplatesResult shape has no defined error codes (in
+	// fact, none of the EC2 API shapes do). We will need to create exceptions
+	// config in the generate.yaml for EC2, but this will take quite some
+	// manual work. For now, return UNKNOWN
+	assert.Equal("UNKNOWN", crd.ExceptionCode(404))
+
+	specFields := crd.SpecFields
+	statusFields := crd.StatusFields
+
+	expSpecFieldCamel := []string{
+		// TODO(jaypipes): DryRun and ClientToken are examples of two fields in
+		// the resource input shape that need to be stripped out of the CRD. We
+		// need to instruct the code generator that these types of fields are
+		// not germane to the resource itself...
+		"ClientToken",
+		"DryRun",
+		"LaunchTemplateData",
+		"LaunchTemplateName",
+		// TODO(jaypipes): Here's an example of where we need to instruct the
+		// code generator to rename the "TagSpecifications" field to simply
+		// "Tags" and place it into the common Spec.Tags field.
+		"TagSpecifications",
+		"VersionDescription",
+	}
+	assert.Equal(expSpecFieldCamel, attrCamelNames(specFields))
+
+	expStatusFieldCamel := []string{
+		"CreateTime",
+		"CreatedBy",
+		"DefaultVersionNumber",
+		"LatestVersionNumber",
+		// TODO(jaypipes): Handle "Id" Fields like "LaunchTemplateId" in the
+		// same way as we handle ARN-ified modern service APIs and use the
+		// SDKMapper to instruct the code generator that this field represents
+		// the primary resource object's identifier field.
+		"LaunchTemplateID",
+		// LaunchTemplateName excluded because it matches input shape.,
+		// TODO(jaypipes): Tags field should be excluded because it is the same
+		// as the input shape's "TagSpecifications" field...
+		"Tags",
+	}
+	assert.Equal(expStatusFieldCamel, attrCamelNames(statusFields))
+
+	// LaunchTemplateName is in the LaunchTemplate resource's CreateTopicInput shape and also
+	// returned in the CreateLaunchTemplateResult shape, so it should have
+	// Go code to set the Input Shape member from the Spec field but not set a
+	// Status field from the Create Output Shape member
+	expCreateInput := `
+	if r.ko.Spec.ClientToken != nil {
+		res.SetClientToken(*r.ko.Spec.ClientToken)
+	}
+	if r.ko.Spec.DryRun != nil {
+		res.SetDryRun(*r.ko.Spec.DryRun)
+	}
+	if r.ko.Spec.LaunchTemplateData != nil {
+		f2 := &svcsdk.RequestLaunchTemplateData{}
+		if r.ko.Spec.LaunchTemplateData.BlockDeviceMappings != nil {
+			f2f0 := []*svcsdk.LaunchTemplateBlockDeviceMappingRequest{}
+			for _, f2f0iter := range r.ko.Spec.LaunchTemplateData.BlockDeviceMappings {
+				f2f0elem := &svcsdk.LaunchTemplateBlockDeviceMappingRequest{}
+				if f2f0iter.DeviceName != nil {
+					f2f0elem.SetDeviceName(*f2f0iter.DeviceName)
+				}
+				if f2f0iter.EBS != nil {
+					f2f0elemf1 := &svcsdk.LaunchTemplateEbsBlockDeviceRequest{}
+					if f2f0iter.EBS.DeleteOnTermination != nil {
+						f2f0elemf1.SetDeleteOnTermination(*f2f0iter.EBS.DeleteOnTermination)
+					}
+					if f2f0iter.EBS.Encrypted != nil {
+						f2f0elemf1.SetEncrypted(*f2f0iter.EBS.Encrypted)
+					}
+					if f2f0iter.EBS.IOPS != nil {
+						f2f0elemf1.SetIops(*f2f0iter.EBS.IOPS)
+					}
+					if f2f0iter.EBS.KMSKeyID != nil {
+						f2f0elemf1.SetKmsKeyId(*f2f0iter.EBS.KMSKeyID)
+					}
+					if f2f0iter.EBS.SnapshotID != nil {
+						f2f0elemf1.SetSnapshotId(*f2f0iter.EBS.SnapshotID)
+					}
+					if f2f0iter.EBS.VolumeSize != nil {
+						f2f0elemf1.SetVolumeSize(*f2f0iter.EBS.VolumeSize)
+					}
+					if f2f0iter.EBS.VolumeType != nil {
+						f2f0elemf1.SetVolumeType(*f2f0iter.EBS.VolumeType)
+					}
+					f2f0elem.SetEbs(f2f0elemf1)
+				}
+				if f2f0iter.NoDevice != nil {
+					f2f0elem.SetNoDevice(*f2f0iter.NoDevice)
+				}
+				if f2f0iter.VirtualName != nil {
+					f2f0elem.SetVirtualName(*f2f0iter.VirtualName)
+				}
+				f2f0 = append(f2f0, f2f0elem)
+			}
+			f2.SetBlockDeviceMappings(f2f0)
+		}
+		if r.ko.Spec.LaunchTemplateData.CapacityReservationSpecification != nil {
+			f2f1 := &svcsdk.LaunchTemplateCapacityReservationSpecificationRequest{}
+			if r.ko.Spec.LaunchTemplateData.CapacityReservationSpecification.CapacityReservationPreference != nil {
+				f2f1.SetCapacityReservationPreference(*r.ko.Spec.LaunchTemplateData.CapacityReservationSpecification.CapacityReservationPreference)
+			}
+			if r.ko.Spec.LaunchTemplateData.CapacityReservationSpecification.CapacityReservationTarget != nil {
+				f2f1f1 := &svcsdk.CapacityReservationTarget{}
+				if r.ko.Spec.LaunchTemplateData.CapacityReservationSpecification.CapacityReservationTarget.CapacityReservationID != nil {
+					f2f1f1.SetCapacityReservationId(*r.ko.Spec.LaunchTemplateData.CapacityReservationSpecification.CapacityReservationTarget.CapacityReservationID)
+				}
+				f2f1.SetCapacityReservationTarget(f2f1f1)
+			}
+			f2.SetCapacityReservationSpecification(f2f1)
+		}
+		if r.ko.Spec.LaunchTemplateData.CPUOptions != nil {
+			f2f2 := &svcsdk.LaunchTemplateCpuOptionsRequest{}
+			if r.ko.Spec.LaunchTemplateData.CPUOptions.CoreCount != nil {
+				f2f2.SetCoreCount(*r.ko.Spec.LaunchTemplateData.CPUOptions.CoreCount)
+			}
+			if r.ko.Spec.LaunchTemplateData.CPUOptions.ThreadsPerCore != nil {
+				f2f2.SetThreadsPerCore(*r.ko.Spec.LaunchTemplateData.CPUOptions.ThreadsPerCore)
+			}
+			f2.SetCpuOptions(f2f2)
+		}
+		if r.ko.Spec.LaunchTemplateData.CreditSpecification != nil {
+			f2f3 := &svcsdk.CreditSpecificationRequest{}
+			if r.ko.Spec.LaunchTemplateData.CreditSpecification.CPUCredits != nil {
+				f2f3.SetCpuCredits(*r.ko.Spec.LaunchTemplateData.CreditSpecification.CPUCredits)
+			}
+			f2.SetCreditSpecification(f2f3)
+		}
+		if r.ko.Spec.LaunchTemplateData.DisableAPITermination != nil {
+			f2.SetDisableApiTermination(*r.ko.Spec.LaunchTemplateData.DisableAPITermination)
+		}
+		if r.ko.Spec.LaunchTemplateData.EBSOptimized != nil {
+			f2.SetEbsOptimized(*r.ko.Spec.LaunchTemplateData.EBSOptimized)
+		}
+		if r.ko.Spec.LaunchTemplateData.ElasticGPUSpecifications != nil {
+			f2f6 := []*svcsdk.ElasticGpuSpecification{}
+			for _, f2f6iter := range r.ko.Spec.LaunchTemplateData.ElasticGPUSpecifications {
+				f2f6elem := &svcsdk.ElasticGpuSpecification{}
+				if f2f6iter.Type != nil {
+					f2f6elem.SetType(*f2f6iter.Type)
+				}
+				f2f6 = append(f2f6, f2f6elem)
+			}
+			f2.SetElasticGpuSpecifications(f2f6)
+		}
+		if r.ko.Spec.LaunchTemplateData.ElasticInferenceAccelerators != nil {
+			f2f7 := []*svcsdk.LaunchTemplateElasticInferenceAccelerator{}
+			for _, f2f7iter := range r.ko.Spec.LaunchTemplateData.ElasticInferenceAccelerators {
+				f2f7elem := &svcsdk.LaunchTemplateElasticInferenceAccelerator{}
+				if f2f7iter.Count != nil {
+					f2f7elem.SetCount(*f2f7iter.Count)
+				}
+				if f2f7iter.Type != nil {
+					f2f7elem.SetType(*f2f7iter.Type)
+				}
+				f2f7 = append(f2f7, f2f7elem)
+			}
+			f2.SetElasticInferenceAccelerators(f2f7)
+		}
+		if r.ko.Spec.LaunchTemplateData.HibernationOptions != nil {
+			f2f8 := &svcsdk.LaunchTemplateHibernationOptionsRequest{}
+			if r.ko.Spec.LaunchTemplateData.HibernationOptions.Configured != nil {
+				f2f8.SetConfigured(*r.ko.Spec.LaunchTemplateData.HibernationOptions.Configured)
+			}
+			f2.SetHibernationOptions(f2f8)
+		}
+		if r.ko.Spec.LaunchTemplateData.IAMInstanceProfile != nil {
+			f2f9 := &svcsdk.LaunchTemplateIamInstanceProfileSpecificationRequest{}
+			if r.ko.Spec.LaunchTemplateData.IAMInstanceProfile.ARN != nil {
+				f2f9.SetArn(*r.ko.Spec.LaunchTemplateData.IAMInstanceProfile.ARN)
+			}
+			if r.ko.Spec.LaunchTemplateData.IAMInstanceProfile.Name != nil {
+				f2f9.SetName(*r.ko.Spec.LaunchTemplateData.IAMInstanceProfile.Name)
+			}
+			f2.SetIamInstanceProfile(f2f9)
+		}
+		if r.ko.Spec.LaunchTemplateData.ImageID != nil {
+			f2.SetImageId(*r.ko.Spec.LaunchTemplateData.ImageID)
+		}
+		if r.ko.Spec.LaunchTemplateData.InstanceInitiatedShutdownBehavior != nil {
+			f2.SetInstanceInitiatedShutdownBehavior(*r.ko.Spec.LaunchTemplateData.InstanceInitiatedShutdownBehavior)
+		}
+		if r.ko.Spec.LaunchTemplateData.InstanceMarketOptions != nil {
+			f2f12 := &svcsdk.LaunchTemplateInstanceMarketOptionsRequest{}
+			if r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.MarketType != nil {
+				f2f12.SetMarketType(*r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.MarketType)
+			}
+			if r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.SpotOptions != nil {
+				f2f12f1 := &svcsdk.LaunchTemplateSpotMarketOptionsRequest{}
+				if r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.SpotOptions.BlockDurationMinutes != nil {
+					f2f12f1.SetBlockDurationMinutes(*r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.SpotOptions.BlockDurationMinutes)
+				}
+				if r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.SpotOptions.InstanceInterruptionBehavior != nil {
+					f2f12f1.SetInstanceInterruptionBehavior(*r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.SpotOptions.InstanceInterruptionBehavior)
+				}
+				if r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.SpotOptions.MaxPrice != nil {
+					f2f12f1.SetMaxPrice(*r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.SpotOptions.MaxPrice)
+				}
+				if r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.SpotOptions.SpotInstanceType != nil {
+					f2f12f1.SetSpotInstanceType(*r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.SpotOptions.SpotInstanceType)
+				}
+				if r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.SpotOptions.ValidUntil != nil {
+					f2f12f1.SetValidUntil(r.ko.Spec.LaunchTemplateData.InstanceMarketOptions.SpotOptions.ValidUntil.Time)
+				}
+				f2f12.SetSpotOptions(f2f12f1)
+			}
+			f2.SetInstanceMarketOptions(f2f12)
+		}
+		if r.ko.Spec.LaunchTemplateData.InstanceType != nil {
+			f2.SetInstanceType(*r.ko.Spec.LaunchTemplateData.InstanceType)
+		}
+		if r.ko.Spec.LaunchTemplateData.KernelID != nil {
+			f2.SetKernelId(*r.ko.Spec.LaunchTemplateData.KernelID)
+		}
+		if r.ko.Spec.LaunchTemplateData.KeyName != nil {
+			f2.SetKeyName(*r.ko.Spec.LaunchTemplateData.KeyName)
+		}
+		if r.ko.Spec.LaunchTemplateData.LicenseSpecifications != nil {
+			f2f16 := []*svcsdk.LaunchTemplateLicenseConfigurationRequest{}
+			for _, f2f16iter := range r.ko.Spec.LaunchTemplateData.LicenseSpecifications {
+				f2f16elem := &svcsdk.LaunchTemplateLicenseConfigurationRequest{}
+				if f2f16iter.LicenseConfigurationARN != nil {
+					f2f16elem.SetLicenseConfigurationArn(*f2f16iter.LicenseConfigurationARN)
+				}
+				f2f16 = append(f2f16, f2f16elem)
+			}
+			f2.SetLicenseSpecifications(f2f16)
+		}
+		if r.ko.Spec.LaunchTemplateData.MetadataOptions != nil {
+			f2f17 := &svcsdk.LaunchTemplateInstanceMetadataOptionsRequest{}
+			if r.ko.Spec.LaunchTemplateData.MetadataOptions.HTTPEndpoint != nil {
+				f2f17.SetHttpEndpoint(*r.ko.Spec.LaunchTemplateData.MetadataOptions.HTTPEndpoint)
+			}
+			if r.ko.Spec.LaunchTemplateData.MetadataOptions.HTTPPutResponseHopLimit != nil {
+				f2f17.SetHttpPutResponseHopLimit(*r.ko.Spec.LaunchTemplateData.MetadataOptions.HTTPPutResponseHopLimit)
+			}
+			if r.ko.Spec.LaunchTemplateData.MetadataOptions.HTTPTokens != nil {
+				f2f17.SetHttpTokens(*r.ko.Spec.LaunchTemplateData.MetadataOptions.HTTPTokens)
+			}
+			f2.SetMetadataOptions(f2f17)
+		}
+		if r.ko.Spec.LaunchTemplateData.Monitoring != nil {
+			f2f18 := &svcsdk.LaunchTemplatesMonitoringRequest{}
+			if r.ko.Spec.LaunchTemplateData.Monitoring.Enabled != nil {
+				f2f18.SetEnabled(*r.ko.Spec.LaunchTemplateData.Monitoring.Enabled)
+			}
+			f2.SetMonitoring(f2f18)
+		}
+		if r.ko.Spec.LaunchTemplateData.NetworkInterfaces != nil {
+			f2f19 := []*svcsdk.LaunchTemplateInstanceNetworkInterfaceSpecificationRequest{}
+			for _, f2f19iter := range r.ko.Spec.LaunchTemplateData.NetworkInterfaces {
+				f2f19elem := &svcsdk.LaunchTemplateInstanceNetworkInterfaceSpecificationRequest{}
+				if f2f19iter.AssociatePublicIPAddress != nil {
+					f2f19elem.SetAssociatePublicIpAddress(*f2f19iter.AssociatePublicIPAddress)
+				}
+				if f2f19iter.DeleteOnTermination != nil {
+					f2f19elem.SetDeleteOnTermination(*f2f19iter.DeleteOnTermination)
+				}
+				if f2f19iter.Description != nil {
+					f2f19elem.SetDescription(*f2f19iter.Description)
+				}
+				if f2f19iter.DeviceIndex != nil {
+					f2f19elem.SetDeviceIndex(*f2f19iter.DeviceIndex)
+				}
+				if f2f19iter.Groups != nil {
+					f2f19elemf4 := []*string{}
+					for _, f2f19elemf4iter := range f2f19iter.Groups {
+						var f2f19elemf4elem string
+						f2f19elemf4elem = *f2f19elemf4iter
+						f2f19elemf4 = append(f2f19elemf4, &f2f19elemf4elem)
+					}
+					f2f19elem.SetGroups(f2f19elemf4)
+				}
+				if f2f19iter.InterfaceType != nil {
+					f2f19elem.SetInterfaceType(*f2f19iter.InterfaceType)
+				}
+				if f2f19iter.IPv6AddressCount != nil {
+					f2f19elem.SetIpv6AddressCount(*f2f19iter.IPv6AddressCount)
+				}
+				if f2f19iter.IPv6Addresses != nil {
+					f2f19elemf7 := []*svcsdk.InstanceIpv6AddressRequest{}
+					for _, f2f19elemf7iter := range f2f19iter.IPv6Addresses {
+						f2f19elemf7elem := &svcsdk.InstanceIpv6AddressRequest{}
+						if f2f19elemf7iter.IPv6Address != nil {
+							f2f19elemf7elem.SetIpv6Address(*f2f19elemf7iter.IPv6Address)
+						}
+						f2f19elemf7 = append(f2f19elemf7, f2f19elemf7elem)
+					}
+					f2f19elem.SetIpv6Addresses(f2f19elemf7)
+				}
+				if f2f19iter.NetworkInterfaceID != nil {
+					f2f19elem.SetNetworkInterfaceId(*f2f19iter.NetworkInterfaceID)
+				}
+				if f2f19iter.PrivateIPAddress != nil {
+					f2f19elem.SetPrivateIpAddress(*f2f19iter.PrivateIPAddress)
+				}
+				if f2f19iter.PrivateIPAddresses != nil {
+					f2f19elemf10 := []*svcsdk.PrivateIpAddressSpecification{}
+					for _, f2f19elemf10iter := range f2f19iter.PrivateIPAddresses {
+						f2f19elemf10elem := &svcsdk.PrivateIpAddressSpecification{}
+						if f2f19elemf10iter.Primary != nil {
+							f2f19elemf10elem.SetPrimary(*f2f19elemf10iter.Primary)
+						}
+						if f2f19elemf10iter.PrivateIPAddress != nil {
+							f2f19elemf10elem.SetPrivateIpAddress(*f2f19elemf10iter.PrivateIPAddress)
+						}
+						f2f19elemf10 = append(f2f19elemf10, f2f19elemf10elem)
+					}
+					f2f19elem.SetPrivateIpAddresses(f2f19elemf10)
+				}
+				if f2f19iter.SecondaryPrivateIPAddressCount != nil {
+					f2f19elem.SetSecondaryPrivateIpAddressCount(*f2f19iter.SecondaryPrivateIPAddressCount)
+				}
+				if f2f19iter.SubnetID != nil {
+					f2f19elem.SetSubnetId(*f2f19iter.SubnetID)
+				}
+				f2f19 = append(f2f19, f2f19elem)
+			}
+			f2.SetNetworkInterfaces(f2f19)
+		}
+		if r.ko.Spec.LaunchTemplateData.Placement != nil {
+			f2f20 := &svcsdk.LaunchTemplatePlacementRequest{}
+			if r.ko.Spec.LaunchTemplateData.Placement.Affinity != nil {
+				f2f20.SetAffinity(*r.ko.Spec.LaunchTemplateData.Placement.Affinity)
+			}
+			if r.ko.Spec.LaunchTemplateData.Placement.AvailabilityZone != nil {
+				f2f20.SetAvailabilityZone(*r.ko.Spec.LaunchTemplateData.Placement.AvailabilityZone)
+			}
+			if r.ko.Spec.LaunchTemplateData.Placement.GroupName != nil {
+				f2f20.SetGroupName(*r.ko.Spec.LaunchTemplateData.Placement.GroupName)
+			}
+			if r.ko.Spec.LaunchTemplateData.Placement.HostID != nil {
+				f2f20.SetHostId(*r.ko.Spec.LaunchTemplateData.Placement.HostID)
+			}
+			if r.ko.Spec.LaunchTemplateData.Placement.HostResourceGroupARN != nil {
+				f2f20.SetHostResourceGroupArn(*r.ko.Spec.LaunchTemplateData.Placement.HostResourceGroupARN)
+			}
+			if r.ko.Spec.LaunchTemplateData.Placement.PartitionNumber != nil {
+				f2f20.SetPartitionNumber(*r.ko.Spec.LaunchTemplateData.Placement.PartitionNumber)
+			}
+			if r.ko.Spec.LaunchTemplateData.Placement.SpreadDomain != nil {
+				f2f20.SetSpreadDomain(*r.ko.Spec.LaunchTemplateData.Placement.SpreadDomain)
+			}
+			if r.ko.Spec.LaunchTemplateData.Placement.Tenancy != nil {
+				f2f20.SetTenancy(*r.ko.Spec.LaunchTemplateData.Placement.Tenancy)
+			}
+			f2.SetPlacement(f2f20)
+		}
+		if r.ko.Spec.LaunchTemplateData.RamDiskID != nil {
+			f2.SetRamDiskId(*r.ko.Spec.LaunchTemplateData.RamDiskID)
+		}
+		if r.ko.Spec.LaunchTemplateData.SecurityGroupIDs != nil {
+			f2f22 := []*string{}
+			for _, f2f22iter := range r.ko.Spec.LaunchTemplateData.SecurityGroupIDs {
+				var f2f22elem string
+				f2f22elem = *f2f22iter
+				f2f22 = append(f2f22, &f2f22elem)
+			}
+			f2.SetSecurityGroupIds(f2f22)
+		}
+		if r.ko.Spec.LaunchTemplateData.SecurityGroups != nil {
+			f2f23 := []*string{}
+			for _, f2f23iter := range r.ko.Spec.LaunchTemplateData.SecurityGroups {
+				var f2f23elem string
+				f2f23elem = *f2f23iter
+				f2f23 = append(f2f23, &f2f23elem)
+			}
+			f2.SetSecurityGroups(f2f23)
+		}
+		if r.ko.Spec.LaunchTemplateData.TagSpecifications != nil {
+			f2f24 := []*svcsdk.LaunchTemplateTagSpecificationRequest{}
+			for _, f2f24iter := range r.ko.Spec.LaunchTemplateData.TagSpecifications {
+				f2f24elem := &svcsdk.LaunchTemplateTagSpecificationRequest{}
+				if f2f24iter.ResourceType != nil {
+					f2f24elem.SetResourceType(*f2f24iter.ResourceType)
+				}
+				if f2f24iter.Tags != nil {
+					f2f24elemf1 := []*svcsdk.Tag{}
+					for _, f2f24elemf1iter := range f2f24iter.Tags {
+						f2f24elemf1elem := &svcsdk.Tag{}
+						if f2f24elemf1iter.Key != nil {
+							f2f24elemf1elem.SetKey(*f2f24elemf1iter.Key)
+						}
+						if f2f24elemf1iter.Value != nil {
+							f2f24elemf1elem.SetValue(*f2f24elemf1iter.Value)
+						}
+						f2f24elemf1 = append(f2f24elemf1, f2f24elemf1elem)
+					}
+					f2f24elem.SetTags(f2f24elemf1)
+				}
+				f2f24 = append(f2f24, f2f24elem)
+			}
+			f2.SetTagSpecifications(f2f24)
+		}
+		if r.ko.Spec.LaunchTemplateData.UserData != nil {
+			f2.SetUserData(*r.ko.Spec.LaunchTemplateData.UserData)
+		}
+		res.SetLaunchTemplateData(f2)
+	}
+	if r.ko.Spec.LaunchTemplateName != nil {
+		res.SetLaunchTemplateName(*r.ko.Spec.LaunchTemplateName)
+	}
+	if r.ko.Spec.TagSpecifications != nil {
+		f4 := []*svcsdk.TagSpecification{}
+		for _, f4iter := range r.ko.Spec.TagSpecifications {
+			f4elem := &svcsdk.TagSpecification{}
+			if f4iter.ResourceType != nil {
+				f4elem.SetResourceType(*f4iter.ResourceType)
+			}
+			if f4iter.Tags != nil {
+				f4elemf1 := []*svcsdk.Tag{}
+				for _, f4elemf1iter := range f4iter.Tags {
+					f4elemf1elem := &svcsdk.Tag{}
+					if f4elemf1iter.Key != nil {
+						f4elemf1elem.SetKey(*f4elemf1iter.Key)
+					}
+					if f4elemf1iter.Value != nil {
+						f4elemf1elem.SetValue(*f4elemf1iter.Value)
+					}
+					f4elemf1 = append(f4elemf1, f4elemf1elem)
+				}
+				f4elem.SetTags(f4elemf1)
+			}
+			f4 = append(f4, f4elem)
+		}
+		res.SetTagSpecifications(f4)
+	}
+	if r.ko.Spec.VersionDescription != nil {
+		res.SetVersionDescription(*r.ko.Spec.VersionDescription)
+	}
+`
+	assert.Equal(expCreateInput, crd.GoCodeSetInput(model.OpTypeCreate, "r.ko", "res", 1))
+
+	// Check that we properly determined how to find the CreatedBy attribute
+	// within the CreateLaunchTemplateResult shape, which has a single field called
+	// "LaunchTemplate" that contains the CreatedBy field.
+	expCreateOutput := `
+	if resp.LaunchTemplate.CreateTime != nil {
+		ko.Status.CreateTime = &metav1.Time{*resp.LaunchTemplate.CreateTime}
+	}
+	if resp.LaunchTemplate.CreatedBy != nil {
+		ko.Status.CreatedBy = resp.LaunchTemplate.CreatedBy
+	}
+	if resp.LaunchTemplate.DefaultVersionNumber != nil {
+		ko.Status.DefaultVersionNumber = resp.LaunchTemplate.DefaultVersionNumber
+	}
+	if resp.LaunchTemplate.LatestVersionNumber != nil {
+		ko.Status.LatestVersionNumber = resp.LaunchTemplate.LatestVersionNumber
+	}
+	if resp.LaunchTemplate.LaunchTemplateId != nil {
+		ko.Status.LaunchTemplateID = resp.LaunchTemplate.LaunchTemplateId
+	}
+	if resp.LaunchTemplate.Tags != nil {
+		f6 := []*svcapitypes.Tag{}
+		for _, f6iter := range resp.LaunchTemplate.Tags {
+			f6elem := &svcapitypes.Tag{}
+			if f6iter.Key != nil {
+				f6elem.Key = f6iter.Key
+			}
+			if f6iter.Value != nil {
+				f6elem.Value = f6iter.Value
+			}
+			f6 = append(f6, f6elem)
+		}
+		ko.Status.Tags = f6
+	}
+`
+	assert.Equal(expCreateOutput, crd.GoCodeSetOutput(model.OpTypeCreate, "resp", "ko.Status", 1))
+
+	// The EC2 LaunchTemplate API has a "normal" set of CUD operations:
+	//
+	// * CreateLaunchTemplate
+	// * ModifyLaunchTemplate
+	// * DeleteLaunchTemplate
+	require.NotNil(crd.Ops)
+
+	assert.NotNil(crd.Ops.Create)
+	assert.NotNil(crd.Ops.Delete)
+	assert.NotNil(crd.Ops.Update)
+
+	// However, oddly, there is no ReadOne operation. There is only a
+	// ReadMany/List operation "DescribeLaunchTemplates" :(
+	//
+	// TODO(jaypipes): Develop strategy for informing the code generator via
+	// the SDKMapper that certain APIs don't have ReadOne but only ReadMany
+	// APIs...
+	assert.Nil(crd.Ops.ReadOne)
+	assert.NotNil(crd.Ops.ReadMany)
+}
