+++ pkg/model/ecr_test.go
@@ -1,0 +1,238 @@
+// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"). You may
+// not use this file except in compliance with the License. A copy of the
+// License is located at
+//
+//	 http://aws.amazon.com/apache2.0/
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package model_test
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/aws/aws-controllers-k8s/pkg/model"
+	"github.com/aws/aws-controllers-k8s/pkg/testutil"
+)
+
+func TestECRRepository(t *testing.T) {
+	assert := assert.New(t)
+	require := require.New(t)
+
+	sh := testutil.NewSchemaHelperForService(t, "ecr")
+
+	crds, err := sh.GetCRDs()
+	require.Nil(err)
+
+	crd := getCRDByName("Repository", crds)
+	require.NotNil(crd)
+
+	// The ECR Repository API has just the C and R of the normal CRUD
+	// operations:
+	//
+	// * CreateRepository
+	// * DeleteRepository
+	require.NotNil(crd.Ops)
+
+	assert.NotNil(crd.Ops.Create)
+	assert.NotNil(crd.Ops.Delete)
+
+	// There is no DescribeRepository operation. There is a List operation for
+	// Repositories, though: DescribeRepositories
+	assert.Nil(crd.Ops.ReadOne)
+	assert.NotNil(crd.Ops.ReadMany)
+
+	// There is no update operation (you need to call various SetXXX operations
+	// on the Repository's components
+	assert.Nil(crd.Ops.Update)
+
+	// The DescribeRepositories operation has the following definition:
+	//
+	//    "DescribeRepositories":{
+	//      "name":"DescribeRepositories",
+	//      "http":{
+	//        "method":"POST",
+	//        "requestUri":"/"
+	//      },
+	//      "input":{"shape":"DescribeRepositoriesRequest"},
+	//      "output":{"shape":"DescribeRepositoriesResponse"},
+	//      "errors":[
+	//        {"shape":"ServerException"},
+	//        {"shape":"InvalidParameterException"},
+	//        {"shape":"RepositoryNotFoundException"}
+	//      ]
+	//    },
+	//
+	// NOTE: This is UNUSUAL for a List operation to return a 404 Not Found.
+	// Typically a return of zero results for a List operation results in a 200
+	// OK.
+	//
+	// Where the RepositoryNotFoundException shape looks like this:
+	//
+	//    "RepositoryNotFoundException":{
+	//      "type":"structure",
+	//      "members":{
+	//        "message":{"shape":"ExceptionMessage"}
+	//      },
+	//      "exception":true
+	//    },
+	//
+	// Which does not indicate that the error is a 404 :( So, the logic in the
+	// CRD.ExceptionCode(404) method needs to get its override from the
+	// generate.yaml configuration file.
+	assert.Equal("RepositoryNotFoundException", crd.ExceptionCode(404))
+
+	specFields := crd.SpecFields
+	statusFields := crd.StatusFields
+
+	// The ECR API uses a REST-like API that uses "wrapper" single-member
+	// objects in the JSON response for the create/describe calls. In other
+	// words, the returned result from the CreateRepository API looks like
+	// this:
+	//
+	// {
+	//   "repository": {
+	//	 .. bunch of fields for the repository ..
+	//   }
+	// }
+	//
+	// This test is verifying that we're properly "unwrapping" the structs and
+	// putting the repository object's fields into the Spec and Status for the
+	// Repository CRD.
+	expSpecFieldCamel := []string{
+		"ImageScanningConfiguration",
+		"ImageTagMutability",
+		"RepositoryName",
+		"Tags",
+	}
+	assert.Equal(expSpecFieldCamel, attrCamelNames(specFields))
+
+	// ImageScanningConfiguration is in the Repository resource's
+	// CreateRepositoryInput shape and also returned in the
+	// CreateRepositoryOutput shape, so it should produce Go code to set the
+	// appropriate input shape member.
+	expCreateInput := `
+	if r.ko.Spec.ImageScanningConfiguration != nil {
+		f0 := &svcsdk.ImageScanningConfiguration{}
+		if r.ko.Spec.ImageScanningConfiguration.ScanOnPush != nil {
+			f0.SetScanOnPush(*r.ko.Spec.ImageScanningConfiguration.ScanOnPush)
+		}
+		res.SetImageScanningConfiguration(f0)
+	}
+	if r.ko.Spec.ImageTagMutability != nil {
+		res.SetImageTagMutability(*r.ko.Spec.ImageTagMutability)
+	}
+	if r.ko.Spec.RepositoryName != nil {
+		res.SetRepositoryName(*r.ko.Spec.RepositoryName)
+	}
+	if r.ko.Spec.Tags != nil {
+		f3 := []*svcsdk.Tag{}
+		for _, f3iter := range r.ko.Spec.Tags {
+			f3elem := &svcsdk.Tag{}
+			if f3iter.Key != nil {
+				f3elem.SetKey(*f3iter.Key)
+			}
+			if f3iter.Value != nil {
+				f3elem.SetValue(*f3iter.Value)
+			}
+			f3 = append(f3, f3elem)
+		}
+		res.SetTags(f3)
+	}
+`
+	assert.Equal(expCreateInput, crd.GoCodeSetInput(model.OpTypeCreate, "r.ko", "res", 1))
+
+	expStatusFieldCamel := []string{
+		"CreatedAt",
+		// "ImageScanningConfiguration" removed because it is contained in the
+		// input shape and therefore exists in the Spec
+		// "ImageTagMutability" removed because it is contained in the input
+		// shape and therefore exists in the Spec
+		"RegistryID",
+		// "RepositoryARN" removed because it refers to the primary object's
+		// ARN and the SDKMapper identified it for mapping to the standard
+		// Status.ACKResourceMetadata.ARN field
+		// "RepositoryName" removed because it is contained in the input shape
+		// and therefore exists in the Spec
+		"RepositoryURI",
+	}
+	assert.Equal(expStatusFieldCamel, attrCamelNames(statusFields))
+
+	// Check that we properly determined how to find the RegistryID attribute
+	// within the CreateRepositoryOutput shape, which has a single field called
+	// "Repository" that contains the RegistryId field.
+	expCreateOutput := `
+	if resp.Repository.CreatedAt != nil {
+		ko.Status.CreatedAt = &metav1.Time{*resp.Repository.CreatedAt}
+	}
+	if resp.Repository.RegistryId != nil {
+		ko.Status.RegistryID = resp.Repository.RegistryId
+	}
+	if resp.Repository.RepositoryUri != nil {
+		ko.Status.RepositoryURI = resp.Repository.RepositoryUri
+	}
+`
+	assert.Equal(expCreateOutput, crd.GoCodeSetOutput(model.OpTypeCreate, "resp", "ko.Status", 1))
+
+	// Check that the DescribeRepositories output is filtered by the
+	// RepositoryName field of the CR's Spec, since there is no ReadOne
+	// operation for ECR and we have no yet implemented a heuristic that allows
+	// the DescribeRepositoriesInput.RepositoryNames filter from the CR's
+	// Spec.RepositoryName field.
+	expReadManyOutput := `
+	if len(resp.Repositories) == 0 {
+		return nil, ackerr.NotFound
+	}
+	found := false
+	for _, elem := range resp.Repositories {
+		if elem.CreatedAt != nil {
+			ko.Status.CreatedAt = &metav1.Time{*elem.CreatedAt}
+		}
+		if elem.ImageScanningConfiguration != nil {
+			f1 := &svcapitypes.ImageScanningConfiguration{}
+			if elem.ImageScanningConfiguration.ScanOnPush != nil {
+				f1.ScanOnPush = elem.ImageScanningConfiguration.ScanOnPush
+			}
+			ko.Spec.ImageScanningConfiguration = f1
+		}
+		if elem.ImageTagMutability != nil {
+			ko.Spec.ImageTagMutability = elem.ImageTagMutability
+		}
+		if elem.RegistryId != nil {
+			ko.Status.RegistryID = elem.RegistryId
+		}
+		if elem.RepositoryArn != nil {
+			if ko.Status.ACKResourceMetadata == nil {
+				ko.Status.ACKResourceMetadata = &ackv1alpha1.ResourceMetadata{}
+			}
+			tmpARN := ackv1alpha1.AWSResourceName(*elem.RepositoryArn)
+			ko.Status.ACKResourceMetadata.ARN = &tmpARN
+		}
+		if elem.RepositoryName != nil {
+			if ko.Spec.RepositoryName != nil {
+				if *elem.RepositoryName != *ko.Spec.RepositoryName {
+					continue
+				}
+			}
+			ko.Spec.RepositoryName = elem.RepositoryName
+		}
+		if elem.RepositoryUri != nil {
+			ko.Status.RepositoryURI = elem.RepositoryUri
+		}
+		found = true
+		break
+	}
+	if !found {
+		return nil, ackerr.NotFound
+	}
+`
+	assert.Equal(expReadManyOutput, crd.GoCodeSetOutput(model.OpTypeList, "resp", "ko", 1))
+}
